KL-163/refactor: separate CommentService
